+++ mdcx/controllers/main_window/load_config.py
@@ -28,7 +28,7 @@ from mdcx.config.enums import (
 from mdcx.config.extend import get_movie_path_setting
 from mdcx.config.manager import manager
 from mdcx.config.resources import resources
-from mdcx.consts import GITHUB_ISSUES_URL, IS_WINDOWS, LOCAL_VERSION, SYSTEM_INFO, VERSION_NAME
+from mdcx.consts import GITHUB_ISSUES_URL, IS_WINDOWS, LOCAL_VERSION, MAIN_PATH, SYSTEM_INFO, VERSION_NAME
 from mdcx.gen.field_enums import CrawlerResultFields
 from mdcx.models.flags import Flags
 from mdcx.signals import signal_qt
@@ -1086,12 +1086,12 @@ def load_config(self: "MyMAinWindow"):
                     )
             if Switch.PASSTHROUGH in switch_on:
                 self.Ui.checkBox_highdpi_passthrough.setChecked(True)
-                if not os.path.isfile("highdpi_passthrough"):
-                    open("highdpi_passthrough", "w").close()
+                if not os.path.isfile(os.path.join(MAIN_PATH, "highdpi_passthrough")):
+                    open(os.path.join(MAIN_PATH, "highdpi_passthrough"), "w").close()
             else:
                 self.Ui.checkBox_highdpi_passthrough.setChecked(False)
-                if os.path.isfile("highdpi_passthrough"):
-                    delete_file_sync("highdpi_passthrough")
+                if os.path.isfile(os.path.join(MAIN_PATH, "highdpi_passthrough")):
+                    delete_file_sync(os.path.join(MAIN_PATH, "highdpi_passthrough"))
         else:
             self.Ui.checkBox_highdpi_passthrough.setEnabled(False)